feat: statusv1 implementation
This implements porcelain=v1 parsing in a fairly robust fashion,
with similar architecture to the statusv2 parser.

Despite its many warts and inconsistencies, the porcelain=v1 format
is a considerably simpler format (providing much less data), and remains
the default when `--porcelain` is passed to git without an argument for
reverse compatibility.

Notably, while I parse and extract the branch name, upstream name, and
ahead/behind info from the --branch header in my usage of porcelain=v1
in github.com/mroth/scmpuff, it does not appear to be a part of the
formal specification for porcelain=v1 which is guaranteed to not change,
and is not documented anywhere in git docs, so I will likely leave that
for application level code and omit from this library, and only provide
the raw header data here to keep this as a pure parser. Users should
prefer porcelain=v2 if they need branch data as it is provided in the
specification.